+++ src/pryngles/optics.py
@@ -6,9 +6,25 @@
 #.##......##..##....##....##..##..##..##..##......##..........##.#
 #.##......##..##....##....##..##...####...######..######...####..#
 #................................................................#
-
+#                                                                #
 # PlanetaRY spanGLES                                             #
 #                                                                #
 ##################################################################
 # License http://github.com/seap-udea/pryngles-public            #
 ##################################################################
+#!/usr/bin/env python
+# coding: utf-8
+
+# # Pryngles module: optics
+# 
+# Template of a module
+
+# ## External modules
+
+#@external
+from pryngles import *
+#@end:external
+
+Scatterer_doc="""This is the basic class of a scatterer
+"""
+